v0.054(O) - Optimized
We are working on "modes" that allow users to simply "click" and apply our extremely complex math.

I'm trying to name them simple things while also providing enough details to be useful.

Also, this is personally for me so I'm actually trying to "design" the idea now.

Its going to take a few iterations.

Everything should be working, as it actually is now on my end. Good luck.